+++ src/api/index.ts
@@ -1,4 +1,4 @@
 export * from './axios';
 export * from './mocks';
-export * from './models';
+export * from './types';
 export * from './hooks';
